+++ buildSrc/CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2023-05-29
+
+### Added
+
+- Support for the Maven snapshot repository if enabled by the build
+  configuration.
+
+### Changed
+
+- Make the use of Maven Local determined by the build configuration.
+
+### Fixed
+
+- Fix the name of the "publish to maven central" targets in GitHub Actions.
+- Fix building for Kotlin/JVM when the Java version is different.
+
 ## [2.0.0] - 2023-05-22
 
 ### Removed
@@ -80,6 +96,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/rhdunn/kotlin-multilpatform-template/compare/2.0.0...HEAD
 [2.0.0]: https://github.com/rhdunn/kotlin-multilpatform-template/compare/1.2.1...2.0.0
+[1.3.0]: https://github.com/rhdunn/kotlin-multilpatform-template/compare/1.2.1...1.3.0
 [1.2.1]: https://github.com/rhdunn/kotlin-multilpatform-template/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/rhdunn/kotlin-multilpatform-template/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/rhdunn/kotlin-multilpatform-template/compare/1.0.0...1.1.0